% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{check_colour_names}
\alias{check_colour_names}
\title{Abort if provided with invalid colour names}
\usage{
check_colour_names(colour_names)
}
\arguments{
\item{colour_names}{A character vector of colour name(s).}
}
\description{
\code{check_colour_names()} will throw an informative error if provided with
colour names that do not exist in the brand colour list.
}
