% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colourdata}
\alias{colourdata}
\title{Colours provided by \code{doestyle}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 11 columns.
}
\source{
\itemize{
\item NSW Design System
(https://github.com/digitalnsw/nsw-design-system/)
\item Department of Education Brand guidelines document
(https://brand.education.nsw.gov.au/)
}
}
\usage{
colourdata
}
\description{
A dataset containing names, hex and RGB values, and other identifiers for
each colour provided in \code{doestyle}. Its content can be visualised with
\code{\link[=show_colours]{show_colours()}} or further manipulated.
}
\seealso{
\code{\link[=show_colours]{show_colours()}}
}
\keyword{datasets}
