% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_math.R
\name{contrast_ratio}
\alias{contrast_ratio}
\title{Compute the contrast ratio between pairs of colours}
\usage{
contrast_ratio(colour1, colour2)
}
\arguments{
\item{colour1, colour2}{Character vectors of hexadecimal colour values.}
}
\value{
\code{contrast_ratio()} returns a numeric vector of contrast ratios, of
the same length as \code{colour1} and \code{colour2}.
}
\description{
Compute contrast ratios between pairs of colours, using the WCAG 2.1 formula.
}
\details{
Contrast ratio is \href{https://www.w3.org/TR/WCAG21/#dfn-contrast-ratio}{defined}
as \eqn{\frac{L_1 + 0.05}{L_2 + 0.05}} where \eqn{L_1} and \eqn{L_2} are the
\href{https://www.w3.org/TR/WCAG21/#dfn-relative-luminance}{relative luminances}
of the lighter and darker colours being compared.

Contrast ratios range from 1 (no colour contrast) to 21 (maximum contrast).
To meet \href{https://www.w3.org/TR/WCAG21/#contrast-minimum}{WCAG 2.1 \emph{minimum} (AA) criteria for accessibility}, the contrast
ratios between adjacent colours must be:
\itemize{
\item For text, at least 4.5
\item For large text (18pt or larger, or 14pt bold), at least 3
\item For non-text plot elements (e.g. patterns, lines, points, borders etc.), at least 3
}

To meet WCAG 2.1 enhanced (AAA) contrast requirements, generally a ratio
above 7 is required.

See WCAG 2.1 success criteria \href{https://www.w3.org/TR/WCAG21/#contrast-minimum}{1.4.3 Contrast (Minimum)}, \href{https://www.w3.org/TR/WCAG21/#contrast-enhanced}{1.4.6 Contrast (Enhanced)} and \href{https://www.w3.org/TR/WCAG21/#non-text-contrast}{1.4.11 Non-text Contrast}.
}
\examples{
backgrounds <- doe_cols("white", "black", "blue-01", "grey-04", "black")
foregrounds <- doe_cols("black", "white", "blue-03", "grey-03", "black")

contrast_ratio(colour1 = backgrounds, colour2 = foregrounds)
}
