% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{doe_colours}
\alias{doe_colours}
\title{Wrapper for \code{palettes::pal_colour()}}
\usage{
doe_colours(colour_names)
}
\arguments{
\item{colour_names}{A character vector of NSW colour names.}
}
\value{
A character vector of colour hex values.
}
\description{
Wraps the colour-creating function from package \code{palettes} so that it accepts
NSW colour names instead of R colour names.
}
