% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{doe_cols}
\alias{doe_cols}
\title{Function to extract doe colors as hex codes}
\usage{
doe_cols(...)
}
\arguments{
\item{...}{Character names of doe_colors}
}
\description{
Function to extract doe colors as hex codes
}
