% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{doe_pal}
\alias{doe_pal}
\title{Return function to interpolate a doe color palette
Makes it possible to create shades between our original colours}
\usage{
doe_pal(palette = "default", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in doe_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a doe color palette
Makes it possible to create shades between our original colours
}
