% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doestyle-package.R
\docType{package}
\name{doestyle-package}
\alias{doestyle}
\alias{doestyle-package}
\title{doestyle: NSW Department of Education branding for figures and tables}
\description{
doestyle is an R package containing functions to help produce brand-compliant figures for NSW Department of Education publications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nsw-education.github.io/doestyle/}
  \item Report bugs at \url{https://github.com/nsw-education/doestyle/issues}
}

}
\author{
\strong{Maintainer}: Samuel Gardiner \email{samuel.gardiner7@det.nsw.edu.au} (\href{https://orcid.org/0000-0002-6752-6969}{ORCID})

Authors:
\itemize{
  \item Samuel Gardiner \email{samuel.gardiner7@det.nsw.edu.au} (\href{https://orcid.org/0000-0002-6752-6969}{ORCID})
  \item Matthew Finkbeiner (\href{https://orcid.org/0000-0003-0373-8150}{ORCID})
  \item Jonathan McGuire
}

Other contributors:
\itemize{
  \item United States Web Design System (Author of included Public Sans font face) [copyright holder]
  \item Dan Williams (Author of included Public Sans font face) [copyright holder]
  \item Pablo Impallari (Author of included Public Sans font face) [copyright holder]
  \item Rodrigo Fuenzalida (Author of included Public Sans font face) [copyright holder]
}

}
\keyword{internal}
