% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_colours}
\alias{get_colours}
\alias{doe_colours}
\alias{doe_cols}
\title{Get brand colours by name}
\usage{
get_colours(...)

doe_colours(...)

doe_cols(...)
}
\arguments{
\item{...}{Colour name(s). May be supplied as a character vector or as
individual arguments.}
}
\value{
\code{get_colours()} returns a character vector containing the hex values
for each requested brand colour. If the input contained named arguments,
their names will be used for the outputs.
}
\description{
Get brand colours by name
}
\details{
\code{get_colours()} will only return hex values for valid NSW and
Department of Education colour names. To list available colours, use
\code{\link[=list_doestyle_colours]{list_doestyle_colours()}}. To plot available colours, use \code{\link[=show_colours]{show_colours()}}.
}
\examples{
get_colours("blue-01", "red-02", "blue-04")
get_colours(c("red-01", "red-02", "red-03", "red-04"))
}
