% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{is_doestyle_colour_name}
\alias{is_doestyle_colour_name}
\title{Test whether given colour names are in the brand colour list}
\usage{
is_doestyle_colour_name(colour_names)
}
\arguments{
\item{colour_names}{A character vector of colour name(s).}
}
\value{
A logical vector.
}
\description{
Test whether given colour names are in the brand colour list
}
