% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{public_sans_available}
\alias{public_sans_available}
\title{Is Public Sans available?}
\usage{
public_sans_available()
}
\value{
\code{TRUE} if Public Sans available; otherwise \code{FALSE}.
}
\description{
Test whether the Public Sans font family is available by searching the
system, user and local font registries.
}
