% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_schools}
\alias{public_schools}
\title{The New South Wales Education Public Schools Master Dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2219 rows and 45 columns.
}
\source{
\url{https://data.nsw.gov.au/data/dataset/nsw-education-nsw-public-schools-master-dataset}
}
\usage{
public_schools
}
\description{
The master dataset contains comprehensive information for all government schools in NSW. Data items include school locations, latitude and longitude coordinates, school type, student enrolment numbers, electorate information, contact details and more.
This dataset is publicly available through the \href{https://data.nsw.gov.au/data/dataset/nsw-education-nsw-public-schools-master-dataset}{Data NSW website}.
}
\keyword{datasets}
