% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_public_sans}
\alias{register_public_sans}
\title{Register Public Sans in the local font registry}
\usage{
register_public_sans()
}
\value{
\code{register_public_sans()} does not have a return value and is called
for its side effects.
}
\description{
Register \code{doestyle}'s embedded copy of the Public Sans font family for use.
This doesn't install the font, but does make it available within this R
session.
}
