% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_math.R
\name{relative_luminance}
\alias{relative_luminance}
\title{Compute the relative luminance of an sRGB colour}
\usage{
relative_luminance(hex_values)
}
\arguments{
\item{hex_values}{A character vector of hex value(s).}
}
\value{
\code{relative_luminance()} returns a numeric vector of luminance values.
}
\description{
Compute the relative luminance of sRGB colour(s). See \href{https://www.w3.org/TR/WCAG21/#dfn-relative-luminance}{WCAG 2.1 definition: relative luminance}.
}
\examples{
relative_luminance(c("#ffffff", "#000000", "#002664", "#d7153a"))
relative_luminance(doe_cols("white", "black", "blue-01", "red-02"))
}
