% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{require_public_sans}
\alias{require_public_sans}
\title{Ensure Public Sans is available for use}
\usage{
require_public_sans()
}
\value{
\code{require_public_sans()} does not have a return value and is called
for its side effects.
}
\description{
Checks that the Public Sans font family is available for use. If it isn't
installed in the system or user font registries, registers \code{doestyle}'s
embedded copy in the local font registry.
}
