% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_doe}
\alias{scale_colour_doe}
\title{Apply a DoE palette to the colour aesthetic}
\usage{
scale_colour_doe(
  palette = doe_palettes$default,
  discrete = TRUE,
  direction = 1,
  ...
)
}
\arguments{
\item{palette}{An object of class
\code{\link[palettes:pal_palette]{palettes_palette}}. Several useful palettes are
ready to use in the list \code{doe_palettes}, which is included with the
\code{doestyle} package. The default is \code{doe_palettes$qualitative$default},
which uses all of the on-brand colours defined in the NSW Department of
Education Brand Guidelines.}

\item{discrete}{Logical indicating whether the desired scale should be
discrete. The default is \code{TRUE}. If \code{FALSE}, a continuous scale with
interpolated colours will be constructed, instead.}

\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are ordered from first to last as they appear in the palette. If
-1, the order of colours is reversed.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}.}
}
\value{
A scale function that can be added to a
\link[ggplot2:ggplot2-package]{ggplot2} plot.
}
\description{
Apply a DoE palette to the colour aesthetic
}
\examples{
library(dplyr)
library(ggplot2)
library(doestyle)

# Plot growth in primary and secondary schools over time
public_schools |>
 filter(Level_of_schooling \%in\% c("Primary School", "Secondary School")) |>
 arrange(Date_1st_teacher) |>
 mutate(
   count = 1,
   cumulative_schools = cumsum(count),
   .by = Level_of_schooling) |>
 ggplot(aes(x = Date_1st_teacher,
            y = cumulative_schools,
            colour = Level_of_schooling)) +
 geom_step() +
 # Add a default DoE colour scale
 scale_colour_doe()

}
\seealso{
\itemize{
\item \link{doe_palettes} for the palettes available in \code{doestyle}.
\item \code{\link[palettes:scale_colour_palette_d]{palettes::scale_colour_palette_d()}}
\item \code{\link[palettes:scale_colour_palette_c]{palettes::scale_colour_palette_c()}}
}
}
