% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{secondary_colours}
\alias{secondary_colours}
\title{Get secondary colours for named brand colours}
\usage{
secondary_colours(...)
}
\arguments{
\item{...}{Colour name(s). May be supplied as a character vector or as
individual arguments.}
}
\value{
\code{secondary_colours()} returns the hex value of an appropriate
secondary colour for each on-brand colour name supplied in its arguments.
}
\description{
Secondary colours are on-brand colours that can be used for pattern fills or
other visual elements that need contrast against an on-brand fill colour.
Each NSW Design System colour has a corresponding secondary colour available
via \code{secondary_colours()}.
}
\examples{
secondary_colours("red-02", "blue-01", "blue-04")
}
