% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{text_colours}
\alias{text_colours}
\title{Get text colours for named brand colours}
\usage{
text_colours(...)
}
\arguments{
\item{...}{Colour name(s). May be supplied as a character vector or as
individual arguments.}
}
\value{
\code{secondary_colours()} returns the hex value of an appropriate text
colour for each on-brand colour name supplied in its arguments.
}
\description{
Text colours are designed for plotting text over an on-brand fill colour.
Each NSW Design System colour has a corresponding text colour available via
\code{text_colours()}.
}
\examples{
text_colours("red-02", "blue-01", "blue-04")
}
