% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-ggplot2.R
\name{theme_doe}
\alias{theme_doe}
\title{Apply the DoE theme to a ggplot2 plot}
\usage{
theme_doe(base_size = 16, base_family = "Public Sans", ...)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family. Defaults to the NSW Government on-brand
font family, \href{https://github.com/uswds/public-sans}{Public Sans}.}

\item{...}{Other parameters passed to \code{ggplot2::theme_minimal()}.}
}
\description{
Apply a minimalist theme.
}
