% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-flextable.R
\name{theme_doe_flextable}
\alias{theme_doe_flextable}
\alias{apply_DoE_zebra_format}
\title{The DoE flextable theme}
\usage{
theme_doe_flextable(x)

apply_DoE_zebra_format(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply a DoE brand-compliant theme to a flextable object.
}
\examples{
head(mtcars) |>
  flextable::flextable() |>
  theme_doe_flextable()
}
