% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colourdata}
\alias{colourdata}
\title{Colours provided by \code{doestyle}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 13 columns.
}
\source{
\itemize{
\item \href{https://github.com/digitalnsw/nsw-design-system/blob/deab23d17f68c799a669d3ee6d6089d9a8ff8a76/src/core/colour/json/palettes.json}{NSW Design System}
\item Department of Education \href{https://brand.education.nsw.gov.au/}{Brand guidelines}
}
}
\usage{
colourdata
}
\description{
A dataset containing names, hex and RGB values, and other identifiers for
each colour provided in \code{doestyle}. Its content can be visualised with
\code{\link[=show_colours]{show_colours()}} or further manipulated.
}
\seealso{
\code{\link[=show_colours]{show_colours()}}
}
\keyword{datasets}
