% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doe_palettes}
\alias{doe_palettes}
\title{Palettes provided by \code{doestyle}}
\format{
An object of class \code{list} of length 3.
}
\usage{
doe_palettes
}
\description{
A list of palettes, organised by type.
}
\details{
\itemize{
\item Qualitative: designed to show differences between nominal classes
\item Sequential: designed to show differences in ordered data that progress in
value from low to high (or vice versa)
\item Diverging: designed to show differences in ordered data that progress away
from a central value to extremes at either end of a scale
}

Palettes can be accessed by their type and name.
}
\examples{
doe_palettes$qualitative$default
doe_palettes$sequential$blues

}
\keyword{datasets}
