% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doestyle-package.R
\docType{package}
\name{doestyle-package}
\alias{doestyle}
\alias{doestyle-package}
\title{doestyle: Apply Department of Education branding to figures and tables}
\description{
doestyle is an R package containing functions to help produce brand-compliant figures for NSW Department of Education publications.
}
\author{
\strong{Maintainer}: Samuel Gardiner \email{samuel.gardiner7@det.nsw.edu.au} (\href{https://orcid.org/0000-0002-6752-6969}{ORCID})

Authors:
\itemize{
  \item Matthew Finkbeiner \email{matthew.finkbeiner@det.nsw.edu.au}
}

}
\keyword{internal}
