% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doestyle-package.R
\docType{package}
\name{doestyle-package}
\alias{doestyle}
\alias{doestyle-package}
\title{doestyle: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Samuel Gardiner \email{samuel.gardiner7@det.nsw.edu.au} (\href{https://orcid.org/0000-0002-6752-6969}{ORCID})

}
\keyword{internal}
