% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_doe}
\alias{scale_colour_doe}
\title{Apply the DoE colour palette to the colour aesthetic}
\usage{
scale_colour_doe(..., aesthetics = "colour")
}
\arguments{
\item{aesthetics}{}
}
\description{
Apply the DoE colour palette to the colour aesthetic
}
