% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_doe}
\alias{scale_colour_doe}
\title{Colour scale constructor for doe colors}
\usage{
scale_colour_doe(
  palette = c("default", "baby_blue_red", "navy_red", "light_shades", "two_blues",
    "three_color", "red_to_blue"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in doe_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour scale constructor for doe colors
}
