% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_doe}
\alias{scale_fill_doe}
\title{Apply the DoE colour palette to the fill aesthetic}
\usage{
scale_fill_doe(..., aesthetics = "fill")
}
\arguments{
\item{aesthetics}{}
}
\description{
Apply the DoE colour palette to the fill aesthetic
}
