% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_doe}
\alias{theme_doe}
\title{Apply the DoE theme to a ggplot2 plot}
\usage{
theme_doe(base_size = 12, ...)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{...}{arguments passed to theme_minimal}
}
\description{
Apply the DoE theme to a ggplot2 plot
}
