% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-ggplot2.R
\name{theme_doe}
\alias{theme_doe}
\title{Apply the DoE theme to a ggplot2 plot}
\usage{
theme_doe(base_size = 16, base_family = "Public Sans", ...)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family. Defaults to the NSW Government on-brand
font family, \href{https://github.com/uswds/public-sans}{Public Sans}.}

\item{...}{Other parameters passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\description{
Apply a minimalist DoE theme. By default, sets the font family to Public Sans
and enables \code{\link[showtext:showtext_auto]{showtext::showtext_auto()}} to allow easy font display in plots.
}
